+++ core/src/loader/context.rs
