+++ core/src/loader/mod.rs
@@ -2,7 +2,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod context;
 pub mod gnu_units;
 mod load;
+
+pub use context::Context;
 
 pub(crate) use load::load_defs;
