+++ core/src/parsing/datetime.rs
@@ -3,7 +3,7 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use crate::ast::{DatePattern, DateToken};
-use crate::context::Context;
+use crate::loader::Context;
 use crate::number::{Dimension, Number};
 use crate::numeric::Numeric;
 use crate::types::{BigInt, BigRat, GenericDateTime};
