+++ core/src/substance.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::context::Context;
+use crate::loader::Context;
 use crate::number::{Dimension, Number};
 use crate::numeric::{Digits, Numeric};
 use crate::output::{PropertyReply, SubstanceReply};
